Fixed issue with resizer, move rendering to resizer module.
